permissions on frontend

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -2,15 +2,16 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
+const { hasPermission } = require('../utils');
 
-const { transport, makeANiceEmail } = require('../mail')
+const { transport, makeANiceEmail } = require('../mail');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
     if (!ctx.request.userId) {
       throw new Error('You must be logged in to continue!');
     }
-    
+
     const item = await ctx.db.mutation.createItem(
       {
         data: {
@@ -48,7 +49,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
 
-    const item = await ctx.db.query.item({ where }, `{ id title}`);
+    const item = await ctx.db.query.item({ where }, `{ id title user {id}}`);
+
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to delete!");
+    }
 
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -120,8 +130,12 @@ const Mutations = {
       from: 'azreed@gmail.com',
       to: user.email,
       subject: 'Your Password Reset Token',
-      html: makeANiceEmail(`Your Password Reset Token is Here! \n\n <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">Click here to Reset</a>`)
-    })
+      html: makeANiceEmail(
+        `Your Password Reset Token is Here! \n\n <a href="${
+          process.env.FRONTEND_URL
+        }/reset?resetToken=${resetToken}">Click here to Reset</a>`
+      )
+    });
 
     return { message: 'Thanks!' };
   },
@@ -156,6 +170,37 @@ const Mutations = {
       maxAge: 1000 * 60 * 60 * 24 * 365
     });
     return updatedUser;
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in!');
+    }
+
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+
+    hasPermission(currentUser, ['ADMIN'], 'PERMISSIONUPDATE');
+
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
